local

+++ app/src/main/java/com/example/toppopapp/data/entities/SongArtistDetailsCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(tableName = "artistAndSong", primaryKeys = ["songId", "id"])
 data class SongArtistDetailsCrossRef (
-    val songId: Int,
+    val songId: Long,
     val id: Long,
 )
